Proposed modularity - Creted app.js for express app initialization and setup, created queries.js module for definition of SQL queries
