+++ api/.env
@@ -1,8 +1,8 @@
 DATABASE_URL=postgres://codemaster:chisom1984@localhost:5432/spacemania
 
-DB_USER=codemaster
+DB_USER=postgres
 DB_HOST=localhost
-DB_NAME=spacemania
-DB_PASSWORD=chisom1984
+DB_NAME=postgres
+DB_PASSWORD=oyinkan1
 DB_PORT=5432
 JWT_SECRET=yoursecretkey
